+++ src/utils/use-keyboard-height.ts
@@ -1,0 +1,25 @@
+import {useState, useEffect} from 'react';
+import {Keyboard, KeyboardEvent} from 'react-native';
+
+export default function useKeyboardHeight() {
+  const [keyboardHeight, setKeyboardHeight] = useState(0);
+
+  function onKeyboardWillShow(e: KeyboardEvent) {
+    setKeyboardHeight(e.endCoordinates.height);
+  }
+
+  function onKeyboardWillHide() {
+    setKeyboardHeight(0);
+  }
+
+  useEffect(() => {
+    Keyboard.addListener('keyboardWillShow', onKeyboardWillShow);
+    Keyboard.addListener('keyboardWillHide', onKeyboardWillHide);
+    return () => {
+      Keyboard.removeListener('onKeyboardWillShow', onKeyboardWillShow);
+      Keyboard.removeListener('keyboardWillHide', onKeyboardWillHide);
+    };
+  }, []);
+
+  return keyboardHeight;
+}
